- Reorganized resources directory. - More info in READMEs. - Started trying to create merged-TOPMed-harmonized-data-file.   - Only finding 773 common phv values accross dbgap and raw-to-harm     files. See README in that directory.
