+++ README.md
@@ -15,10 +15,7 @@ Please read the [Collaboration Guide](CONTRIBUTING.md) before contributing.
     * [datamodel](src/bdchm/datamodel) -- generated
       Python datamodel
 * [tests/](tests/) - Python tests
-* [resources/](resources/)
-  * [map-topmed-raw-to-harmonized/](resources/map-topmed-raw-to-harmonized/)
-    * -- script for generating mapping from TopMed raw variables
-      to TopMed harmonized variables
+* [resources/](resources/) -- Scripts, input files, and output files.
 
 ## Developer Documentation
