+++ resources/README.md
@@ -1,0 +1,21 @@
+## Generate mapping from TopMed raw variables to TopMed harmonized variables
+
+### Contents
+
+* [map-topmed-raw-to-harmonized/](map-topmed-raw-to-harmonized/)
+  * [generate_mapping.py](map-topmed-raw-to-harmonized/generate_mapping.py) -- Script
+  * [raw-to-harmonized-topmed-vars.csv](map-topmed-raw-to-harmonized/raw-to-harmonized-topmed-vars.csv)
+    -- Output
+
+* [human-readable-harmonized-variable-documentation/](human-readable-harmonized-variable-documentation/)
+  * [topmed-variable-documentation.md](human-readable-harmonized-variable-documentation/topmed-variable-documentation.md)
+
+* [merged-TOPMed-harmonized-data-file](merged-TOPMed-harmonized-data-file/)
+
+
+### External resources
+
+* [copies-of-certain-files-from-gdrive/](copies-of-certain-files-from-gdrive/)
+* [harmonized-variable-documentation](harmonized-variable-documentation)
+  * Symbolic link to local clone of [UW-GAC/topmed-dcc-harmonized-phenotypes](https://github.com/UW-GAC/topmed-dcc-harmonized-phenotypes/tree/master/harmonized-variable-documentation).
+    Clone that repo in the same directory as this repo, and the symlink should work.
