+++ resources/copies-of-certain-files-from-gdrive/README.md
@@ -1,0 +1,10 @@
+# Copies of Google Drive files needed for work being done in this repository
+
+This is probably dumb, but I'm putting copies in a subdirectory structure mirroring that
+of the GDrive files.
+
+I'll also list here which files are actually copied and when. And I'll make a symlink to each in this directory.
+
+| Copy date  | File path                                                                                                                                                         | Link                                                                         |
+|------------|-------------------------------------------------------------------------------------------------------------------------------------------------------------------|------------------------------------------------------------------------------|
+| 2024-12-19 | Data Management Core (DMC)/Task 3 - Data Standards/BDC Computational Data Model/Data Modeling and Mapping/BDC Variable Lists/dbgap_variables_priority_cohorts.csv | [dbgap_variables_priority_cohorts.csv](dbgap_variables_priority_cohorts.csv) |
