+++ resources/copies-of-certain-files-from-gdrive/dbgap_variables_priority_cohorts.csv
@@ -1,0 +1,1 @@
+Data Management Core (DMC)/Task 3 - Data Standards/BDC Computational Data Model/Data Modeling and Mapping/BDC Variable Lists/dbgap_variables_priority_cohorts.csv
