+++ resources/harmonized-variable-documentation
@@ -1,0 +1,1 @@
+../../topmed-dcc-harmonized-phenotypes/harmonized-variable-documentation
