+++ resources/human-readable-harmonized-variable-documentation/harmonized-variable-documentation
@@ -1,0 +1,1 @@
+../harmonized-variable-documentation
