--- resources/map-topmed-raw-to-harmonized/README.md
@@ -1,12 +1,0 @@
-### Generate mapping from TopMed raw variables to TopMed harmonized variables
-
-#### Files
-
-* [harmonized-variable-documentation](harmonized-variable-documentation)
-  -- symbolic link to local clone of
-  [UW-GAC/topmed-dcc-harmonized-phenotypes](https://github.com/UW-GAC/topmed-dcc-harmonized-phenotypes/tree/master/harmonized-variable-documentation).
-  Clone that repo in the same directory as this repo, and the symbolin
-  link should work.
-* [generate_mapping.py](generate_mapping.py) -- script
-* [raw-to-harmonized-topmed-vars.csv](raw-to-harmonized-topmed-vars.csv)
-  -- output
