--- resources/map-topmed-raw-to-harmonized/harmonized-variable-documentation
@@ -1,1 +1,0 @@
-../../../topmed-dcc-harmonized-phenotypes/harmonized-variable-documentation
